[refactor] DAOs now hold entities pathnames
