+++ src/dao/ThingDao.ts
@@ -2,5 +2,7 @@ import {BaseDao} from "./BaseDao";
 import {Thing} from "../model/Thing";
 
 export class ThingDao extends BaseDao<Thing> {
-
+    getEntityPathname(): string {
+        return "Things";
+    }
 }
