+++ src/service/SensorThingsService.ts
@@ -14,7 +14,7 @@ export class SensorThingsService {
 
     public async create(entity: Entity): Promise<void> {
         const response: AxiosResponse = await axios.post(
-            [this._endpoint.origin, entity.getURLSuffix()].join('/'),
+            [this._endpoint.origin, entity.getDao(this).getEntityPathname()].join('/'),
             entity.toNetworkObject()
         );
         entity.id = response.data['@iot.id'];
@@ -23,7 +23,7 @@ export class SensorThingsService {
 
     public async update(entity: Entity): Promise<void> {
         await axios.patch(
-            [this._endpoint.origin, entity.entityResourcePathname].join('/'),
+            [this._endpoint.origin, entity.entityResourcePathname(this)].join('/'),
             entity.toNetworkObject()
         );
         return;
@@ -31,7 +31,7 @@ export class SensorThingsService {
 
     public async delete(entity: Entity): Promise<void> {
         await axios.delete(
-            [this._endpoint.origin, entity.entityResourcePathname].join('/')
+            [this._endpoint.origin, entity.entityResourcePathname(this)].join('/')
         );
     }
